+++ CHANGELOG.md
@@ -5,5 +5,9 @@ on [Keep a Changelog](https://keepachangelog.com/en/1.1.0), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New preset: `dist/fnm-install-on-engine-change.json`.
+- New preset: `dist/nvm-install-on-nvmrc-change.json`.
+- New preset: `dist/pnpm-install-on-lockfile-change.json`.
 
 [unreleased]: https://github.com/rainstormy/presets-lefthook